requirements.txt
asgiref==3.5.2
Django==4.1
djangorestframework==3.13.1
djangorestframework-jwt==1.11.0
djongo==1.3.6
dnspython==2.2.1
mongoengine==0.24.2
PyJWT==1.7.1
pymongo==3.12.3
pytz==2022.2.1
sqlparse==0.2.4
